requirements.txt
apimas
django-choices
cliff
